feat: rename topic components for consistency and clarity
