+++ src/renderer/src/components/tap-topics/TopicLineItem.vue
